2.4.2 - Updates to Syndication Calculation
The new calculation is unrestricted, but does a better job at judging whether a season is likely up for syndication or not. Of course, there are examples which give ridiculous results, primarily for shows that aired outside the sample dates, but for the most part the calculation seems reasonable. I've also added additional disclaimer explaining how these are inherently predictions and not necessarily an accurate reflection of real world syndication status.